separates library functions from Tool and from Demo code (#48)
* separates library functions from Tool and from Demo code, updates README to reflect changes

* Update server.js
