+++ lti_lib/oidc.js
@@ -1,4 +1,7 @@
 require('dotenv').config();
+const url = require('url');    
+const Database = require('../lti_lib/mongoDB/Database.js');
+const { platformSchema } = require('../lti_lib/register_platform');
 
 /*
 * Validates OIDC login request.  Checkes required parameters are present.
@@ -26,32 +29,56 @@ function is_valid_oidc_login(req) {
 * @returns if valid request, returns properly formated response object
 * @return if invalid request, returns array of errors with the request
 */
-function create_oidc_response(req) {
-  const errors = is_valid_oidc_login(req);
+function create_oidc_response(req, res) {
+  let errors = [];
 
-  if (errors.length === 0 && req.session.platform_DBinfo) {
-    let response = {
-      scope: 'openid',
-      response_type: 'id_token',
-      client_id: req.session.platform_DBinfo.consumerToolClientID,
-      redirect_uri: process.env.REDIRECT_URI,     // TODO: store in DB per Issuer (Consumer)?
-      login_hint: req.body.login_hint,
-      state: create_unique_string(30, true),
-      response_mode: 'form_post',
-      nonce: create_unique_string(25, false),
-      prompt: 'none'
+  //Save the OIDC Login Request to reference later during current session
+  req.session.login_request = req.body;
+
+  Database.Get('platforms', platformSchema, { consumerUrl: req.session.login_request.iss })
+  .then(dbResult => {
+
+    if (dbResult.length === 1) return dbResult[0]
+    else res.send(['Issuer invalid: not registered']);
+  }).then(platform => {
+    //Save the Platform information from the database to reference later during current session
+    req.session.platform_DBinfo = platform;
+
+    errors = is_valid_oidc_login(req);
+
+    if (errors.length === 0 && req.session.platform_DBinfo) {
+      let response = {
+        scope: 'openid',
+        response_type: 'id_token',
+        client_id: req.session.platform_DBinfo.consumerToolClientID,
+        redirect_uri: process.env.REDIRECT_URI,     // TODO: store in DB per Issuer (Consumer)?
+        login_hint: req.body.login_hint,
+        state: create_unique_string(30, true),
+        response_mode: 'form_post',
+        nonce: create_unique_string(25, false),
+        prompt: 'none'
+      }
+      if (req.body.hasOwnProperty('lti_message_hint')) {
+        response = {
+          ...response,
+          lti_message_hint: req.body.lti_message_hint,
+        };
+      }
+      //Save the OIDC Login Response to reference later during current session
+      req.session.login_response = response;
+
+      res.redirect(url.format({
+        pathname: platform.consumerAuthorizationURL, 
+        query: req.session.login_response
+      }));
+    } else if (!req.session.platform_DBinfo) {
+        errors.push('Issuer invalid: not registered');
     }
-    if (req.body.hasOwnProperty('lti_message_hint')) {
-      response = {
-        ...response,
-        lti_message_hint: req.body.lti_message_hint,
-      };
-    }
-    return response;
-  } else if (!req.session.platform_DBinfo) {
-    errors.push('Issuer invalid: not registered');
+  });
+  //errors were found, so return the errors
+  if (errors.length > 0) {
+    res.send('Error with OIDC Login: ' + errors);
   }
-  return errors;
 }
 
 /*
@@ -70,17 +97,4 @@ function create_unique_string(length, signed) {
   return unique_string;
 }
 
-/*
-* Validate that the state sent with an OIDC launch matches the state that was sent in the OIDC response
-* @param req - HTTP OIDC request to launch Tool.
-* @returns - boolean on whether it is valid or not
-*/
-function is_valid_oidc_launch(req) {
-  //TODO: when state is signed before being sent, need to decode before validation
-  if (req.body.state !== req.session.login_response.state) {
-    return false;
-  }
-  return true;
-}
-
-module.exports = { create_oidc_response, is_valid_oidc_launch };
+module.exports = { create_oidc_response };
